Adjustments after code rebase due to classes being moved (and unable to track them properly by git)
