+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/tezos/DatabaseConversions.scala
@@ -1,9 +1,7 @@
 package tech.cryptonomic.conseil.common.tezos
 
 import scala.util.Try
-import com.typesafe.scalalogging.{LazyLogging, Logger}
-import tech.cryptonomic.conseil.common.tezos.TezosTypes._
-import tech.cryptonomic.conseil.common.tezos.FeeOperations._
+import com.typesafe.scalalogging.LazyLogging
 import cats.{Id, Show}
 import cats.implicits._
 import java.sql.Timestamp
@@ -13,7 +11,7 @@ import java.time.format.DateTimeFormatter
 import monocle.Getter
 import io.scalaland.chimney.dsl._
 import tech.cryptonomic.conseil.common.util.Conversion
-import tech.cryptonomic.conseil.common.tezos.TezosTypes.{BakingRights, Contract, EndorsingRights, FetchRights}
+import tech.cryptonomic.conseil.common.tezos.TezosTypes.{BakingRights, EndorsingRights, FetchRights}
 import tech.cryptonomic.conseil.common.tezos.FeeOperations._
 import tech.cryptonomic.conseil.common.tezos.TezosTypes._
 import tech.cryptonomic.conseil.common.tezos.TezosTypes.Voting.Vote
