+++ conseil-common/src/test/scala/tech/cryptonomic/conseil/common/tezos/TezosDatabaseOperationsTest.scala
@@ -12,7 +12,7 @@ import tech.cryptonomic.conseil.common.testkit.InMemoryDatabase
 import tech.cryptonomic.conseil.common.testkit.util.RandomSeed
 import tech.cryptonomic.conseil.common.tezos.michelson.contracts.{TNSContract, TokenContracts}
 import tech.cryptonomic.conseil.common.tezos.FeeOperations.AverageFees
-import tech.cryptonomic.conseil.common.tezos.Tables.{AccountsHistoryRow, AccountsRow, BlocksRow, FeesRow, ProcessedChainEventsRow, profile}
+import tech.cryptonomic.conseil.common.tezos.Tables.{AccountsHistoryRow, AccountsRow, BlocksRow, FeesRow, ProcessedChainEventsRow}
 import tech.cryptonomic.conseil.common.tezos.TezosTypes._
 
 import scala.concurrent.duration._
