+++ conseil-lorre/src/main/resources/logback.xml
@@ -11,17 +11,32 @@
         </encoder>
     </appender>
 
+    <appender name="JSON" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder class="net.logstash.logback.encoder.LogstashEncoder"/>
+    </appender>
 
     <appender name="ASYNC" class="ch.qos.logback.classic.AsyncAppender">
+        <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
+            <level>${ASYNC_LOG_LEVEL:-INFO}</level>
+        </filter>
         <appender-ref ref="STDOUT" />
+    </appender>
+
+    <appender name="ASYNC_JSON" class="ch.qos.logback.classic.AsyncAppender">
+        <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
+            <level>${JSON_LOG_LEVEL:-OFF}</level>
+        </filter>
+        <appender-ref ref="JSON" />
     </appender>
 
     <logger name="com.base22" level="TRACE"/>
     <logger name="slick" level="INFO"/>
     <logger name="tech.cryptonomic.conseil.common.tezos.BigMapsOperations" level="${LORRE_BIG_MAPS_LOG_LEVEL:-OFF}"/>
+    <logger name="tech.cryptonomic.conseil.common.tezos.michelson.contracts.TNSContracts" level="${LORRE_TNS_LOG_LEVEL:-INFO}"/>
 
     <root level="INFO">
         <appender-ref ref="ASYNC" />
+        <appender-ref ref="ASYNC_JSON" />
     </root>
 
 </configuration>
