+++ conseil-lorre/src/test/scala/tech/cryptonomic/conseil/indexer/tezos/TezosDatabaseOperationsTestFixtures.scala
@@ -3,15 +3,15 @@ package tech.cryptonomic.conseil.indexer.tezos
 import java.sql.Timestamp
 import java.time.{Instant, ZonedDateTime}
 
-import scala.util.Random
-import tech.cryptonomic.conseil.common.tezos.Tables.{AccountsRow, BlocksRow, DelegatesRow, OperationGroupsRow}
-import tech.cryptonomic.conseil.common.tezos.TezosTypes._
-import tech.cryptonomic.conseil.common.tezos.FeeOperations.AverageFees
-import tech.cryptonomic.conseil.common.tezos.TezosTypes.Scripted.Contracts
 import monocle.Optional
 import tech.cryptonomic.conseil.common.testkit.util.{RandomGenerationKit, RandomSeed}
-import tech.cryptonomic.conseil.common.tezos.{Tables, TezosTypes}
-import tech.cryptonomic.conseil.common.tezos.TezosOptics
+import tech.cryptonomic.conseil.common.tezos.FeeOperations.AverageFees
+import tech.cryptonomic.conseil.common.tezos.Tables.{AccountsRow, BakersRow, BlocksRow, OperationGroupsRow}
+import tech.cryptonomic.conseil.common.tezos.TezosTypes.Scripted.Contracts
+import tech.cryptonomic.conseil.common.tezos.TezosTypes._
+import tech.cryptonomic.conseil.common.tezos.{Tables, TezosOptics, TezosTypes}
+
+import scala.util.Random
 
 //TODO This class is a duplicate from conseil-common,
 // which will be updated once entire project will be split properly
@@ -68,7 +68,7 @@ trait TezosDataGeneration extends RandomGenerationKit {
         )
     }.toMap
 
-    accounts.taggedWithBlock(blockHash, blockLevel, Some(time), None)
+    accounts.taggedWithBlock(blockHash, blockLevel, Some(time), None, None)
   }
 
   /* randomly generates a number of delegates with associated block data */
@@ -101,7 +101,7 @@ trait TezosDataGeneration extends RandomGenerationKit {
             )
     }.toMap
 
-    delegates.taggedWithBlock(blockHash, blockLevel, Some(Instant.ofEpochSecond(0)), None)
+    delegates.taggedWithBlock(blockHash, blockLevel, Some(Instant.ofEpochSecond(0)), None, None)
   }
 
   /* randomly populate a number of blocks based on a level range */
@@ -379,7 +379,7 @@ trait TezosDataGeneration extends RandomGenerationKit {
   }
 
   /* randomly generates a number of delegate rows for some block */
-  def generateDelegateRows(howMany: Int, block: BlocksRow)(implicit randomSeed: RandomSeed): List[DelegatesRow] = {
+  def generateDelegateRows(howMany: Int, block: BlocksRow)(implicit randomSeed: RandomSeed): List[BakersRow] = {
     require(
       howMany > 0,
       "the test can only generate a positive number of delegates, you asked for a non positive value"
@@ -389,7 +389,7 @@ trait TezosDataGeneration extends RandomGenerationKit {
     val generateHash: Int => String = alphaNumericGenerator(new Random(randomSeed.seed))
 
     List.fill(howMany) {
-      DelegatesRow(
+      BakersRow(
         blockId = block.hash,
         pkh = generateHash(10),
         balance = Some(0),
